chore: Modified `LocalAuthorityNumberOfPlans` API model
