+++ platform/src/apis/Platform.Api.NonFinancial/Features/EducationHealthCarePlans/Models/History.cs
@@ -1,0 +1,10 @@
+// ReSharper disable PropertyCanBeMadeInitOnly.Global
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+namespace Platform.Api.NonFinancial.Features.EducationHealthCarePlans.Models;
+
+public record History<T>
+{
+    public int? StartYear { get; set; }
+    public int? EndYear { get; set; }
+    public T[]? Plans { get; set; } = [];
+}
