+++ platform/src/apis/Platform.Api.NonFinancial/Features/EducationHealthCarePlans/Models/LocalAuthorityNumberOfPlans.cs
@@ -1,20 +1,11 @@
-﻿namespace Platform.Api.NonFinancial.Features.EducationHealthCarePlans.Models;
-
-public record History<T>
-{
-    public int? StartYear { get; set; }
-    public int? EndYear { get; set; }
-    public T[]? Plans { get; set; } = [];
-}
-
-public record LocalAuthorityNumberOfPlansYear : LocalAuthorityNumberOfPlans
-{
-    public int? Year { get; set; }
-}
+// ReSharper disable PropertyCanBeMadeInitOnly.Global
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+namespace Platform.Api.NonFinancial.Features.EducationHealthCarePlans.Models;
 
 public record LocalAuthorityNumberOfPlans
 {
     public string? Code { get; set; }
+    public string? Name { get; set; }
     public decimal? Total { get; set; }
     public decimal? Mainstream { get; set; }
     public decimal? Resourced { get; set; }
