+++ platform/src/apis/Platform.Api.NonFinancial/Features/EducationHealthCarePlans/Models/LocalAuthorityNumberOfPlansYear.cs
@@ -1,0 +1,8 @@
+﻿// ReSharper disable PropertyCanBeMadeInitOnly.Global
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+namespace Platform.Api.NonFinancial.Features.EducationHealthCarePlans.Models;
+
+public record LocalAuthorityNumberOfPlansYear : LocalAuthorityNumberOfPlans
+{
+    public int? Year { get; set; }
+}
