+++ platform/src/apis/Platform.Api.NonFinancial/Features/EducationHealthCarePlans/Services/EducationHealthCarePlansStubService.cs
@@ -37,6 +37,7 @@ public class EducationHealthCarePlansStubService : IEducationHealthCarePlansServ
                 {
                     Year = j,
                     Code = code,
+                    Name = $"Local authority {code}",
                     Mainstream = mainstream,
                     Resourced = resourced,
                     Special = special,
