+++ web/src/Web.App/Controllers/Api/Responses/EducationHealthCarePlansHistoryResponse.cs
@@ -1,3 +1,4 @@
+// ReSharper disable PropertyCanBeMadeInitOnly.Global
 namespace Web.App.Controllers.Api.Responses;
 
 public record EducationHealthCarePlansHistoryResponse
