+++ web/tests/Web.Tests/Controllers/Api/Mappers/WhenEducationHealthCarePlansHistoryResponseMapperMapsWithCode.cs
@@ -48,6 +48,7 @@ public class WhenEducationHealthCarePlansHistoryResponseMapperMapsWithCode
             Assert.Equal(expected.Total, actual.Total);
             Assert.Equal(expected.Mainstream, actual.Mainstream);
             Assert.Equal(expected.Resourced, actual.Resourced);
+            Assert.Equal(expected.Special, actual.Special);
             Assert.Equal(expected.Independent, actual.Independent);
             Assert.Equal(expected.Hospital, actual.Hospital);
             Assert.Equal(expected.Post16, actual.Post16);
